feat(bank): tests
